+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Added progress bar for train (and test) loss evaluation.
+
 ## [0.3.1] - 2024-11-27
 - Removed old parent directory creation logic for checkpoint file
 - Added dropout option to activated block